+++ src/main/java/org/cotato/csquiz/api/member/dto/UpdateActiveMemberToOldMemberRequest.java
@@ -1,4 +1,4 @@
-package org.cotato.csquiz.api.admin.dto;
+package org.cotato.csquiz.api.member.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
