+++ src/main/java/org/cotato/csquiz/api/member/dto/UpdateMemberRoleRequest.java
@@ -1,4 +1,4 @@
-package org.cotato.csquiz.api.admin.dto;
+package org.cotato.csquiz.api.member.dto;
 
 import org.cotato.csquiz.domain.auth.enums.MemberRole;
 import jakarta.validation.constraints.NotNull;
